User Management API in Go
This project is a simple RESTful API built using the Gin web framework in Go. The API interacts with a PostgreSQL database to perform CRUD (Create, Read, Update, Delete) operations on a "users" table. The API includes the following key features:

Status Check Endpoint: Verifies the health of the API and the database connection.
Create User: Allows the creation of a new user in the database after checking that the user's name or email doesn't already exist.
Get All Users: Retrieves a list of all users stored in the database.
Get User by ID: Fetches the details of a specific user by their ID.
Update User: Updates the name and email of an existing user by ID.
Delete User: Deletes a user from the database by ID.